AI_Search_Knowledge_Graph/3_graph_builder/requirements.txt
# Azure SQL Database
pymssql==2.3.1

# Azure CosmosDB Gremlin API
gremlinpython==3.7.2
aiohttp==3.9.1
async-timeout==4.0.3

# Azure AI Search
azure-search-documents==11.4.0
azure-core==1.30.0
